refactor(ff-preview): extract MasterClock into playback/master_clock.rs
